+++ components/core/ui/theme/src/main/kotlin/com/flipperdevices/core/ui/theme/models/FlipperTypography.kt
@@ -16,6 +16,7 @@ data class FlipperTypography(
     val titleEB20: TextStyle,
     val titleEB18: TextStyle,
     val titleB18: TextStyle,
+    val titleSB18: TextStyle,
     val titleR18: TextStyle,
     val titleM18: TextStyle,
 
@@ -26,6 +27,7 @@ data class FlipperTypography(
     val subtitleB10: TextStyle,
     val subtitleM10: TextStyle,
 
+    val bodySB14: TextStyle,
     val bodyR14: TextStyle,
     val bodyM14: TextStyle,
     val bodyR16: TextStyle,